requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-dotenv==1.0.1
pydantic==2.6.4
openai>=1.30.0
